add exercise to workout feature

+++ app/src/main/java/com/example/digitalfit/modelDb/ExerciseWorkoutCrossRef.kt
@@ -5,5 +5,5 @@ import androidx.room.Entity
 @Entity(primaryKeys = ["exerciseId", "workoutId"], tableName = "exercise_workout")
 class ExerciseWorkoutCrossRef (
     val exerciseId: Int,
-    val workoutId: Int
+    val workoutId: Long
 )
